RPi/requirements.txt
pyyaml>=5.4.1
argparse>=1.2.1
rx>=3.2.0